backend/requirements.txt
python-dotenv>=0.19.0
Pillow>=8.0.0
pytesseract>=0.3.8
setuptools>=65.0.0
flask>=3.0.0